Implement 360-degree movement with keyboard, mouse, and touch support
Changes:
- Client: Add multi-input support (WASD/Arrows, mouse, touch)
- Client: Track pressed keys and cursor position
- Client: Calculate direction angle from input
- Server: Change direction type from u8 to f32 (radians)
- Server: Update movement to use cos/sin with angle
- Server: Update bot AI to use angles instead of discrete directions
- Remove DIRECTIONS constant, use radians throughout

+++ client/src/module_bindings/bot_table.ts
@@ -16,9 +16,9 @@ export default __t.row({
   color: __t.string(),
   score: __t.u32(),
   length: __t.u32(),
-  direction: __t.u8(),
+  direction: __t.f32(),
   alive: __t.bool(),
   x: __t.f32(),
   y: __t.f32(),
-  pendingDirection: __t.u8().name("pending_direction"),
+  pendingDirection: __t.f32().name("pending_direction"),
 });
