+++ client/src/module_bindings/change_direction_reducer.ts
@@ -11,5 +11,5 @@ import {
 } from "spacetimedb";
 
 export default {
-  direction: __t.u8(),
+  direction: __t.f32(),
 };
